split mock and real chatModel tests
